replace core manual test
